Update main.js (#3)
Recently GitHub has updated their HTML for the GitHub Squash and Merge button, and was causing this error:
Uncaught TypeError: Cannot read property 'getAttribute' of null
    at scanForSquashAndMergeButtons (main.js:39)
    at poll (main.js:61)

The reason being is that the button was no longer with a type=submit and therefore was a null. It then failed when we tried to getAttribute of this null variable.

I updated the query selector to find the button correctly with the new introduced updates.